Tighten startup runtime wiring and tests
Fix the remaining startup follow-up items that were still valid on this\nbranch, including Unix SIGHUP ownership, cli_only WASM skipping,\nURL sanitization edge cases, and clearer regression coverage.\n\nThis also cleans up the trybuild compile-contract harness naming and\nstartup shims, updates the startup developer documentation spacing,\nand extracts startup WASM and channel test modules to keep file sizes\nand duplication in bounds.
