+++ src/setup/prompts.rs
@@ -503,6 +503,23 @@ mod tests {
         assert_eq!(stdout, b"*");
         Ok(())
     }
+
+    #[test]
+    fn test_apply_secret_input_effect_sequence_snapshot() -> io::Result<()> {
+        let mut stdout = Vec::new();
+        apply_secret_input_effect(&mut stdout, &SecretInputEffect::MaskChar)?;
+        apply_secret_input_effect(&mut stdout, &SecretInputEffect::Backspace)?;
+        apply_secret_input_effect(&mut stdout, &SecretInputEffect::Submit)?;
+        insta::assert_debug_snapshot!(stdout, @r###"
+        [
+            42,
+            8,
+            32,
+            8,
+        ]
+        "###);
+        Ok(())
+    }
 }
 
 #[derive(Debug, PartialEq, Eq)]
