+++ src/startup/run.rs
@@ -38,7 +38,7 @@ pub(crate) async fn run_agent(ctx: GatewayPhaseContext) -> anyhow::Result<()> {
             },
         channels,
         webhook_server,
-        mut loaded_wasm_channel_names,
+        loaded_wasm_channel_names,
         mut wasm_channel_runtime_state,
         #[cfg(unix)]
         http_channel_state,
@@ -53,7 +53,7 @@ pub(crate) async fn run_agent(ctx: GatewayPhaseContext) -> anyhow::Result<()> {
         components: &components,
         extension_manager: &components.extension_manager,
         wasm_channel_runtime_state: &mut wasm_channel_runtime_state,
-        loaded_wasm_channel_names: &mut loaded_wasm_channel_names[..],
+        loaded_wasm_channel_names: &loaded_wasm_channel_names,
         channels: &channels,
         sse_sender: &sse_sender,
         wasm_channel_owner_ids: &config.channels.wasm_channel_owner_ids,
@@ -109,7 +109,7 @@ struct ChannelPreparation<'a> {
     components: &'a AppComponents,
     extension_manager: &'a Option<Arc<ironclaw::extensions::ExtensionManager>>,
     wasm_channel_runtime_state: &'a mut Option<crate::startup::wasm::WasmChannelRuntimeState>,
-    loaded_wasm_channel_names: &'a mut [String],
+    loaded_wasm_channel_names: &'a [String],
     channels: &'a Arc<ironclaw::channels::ChannelManager>,
     sse_sender: &'a Option<tokio::sync::broadcast::Sender<SseEvent>>,
     wasm_channel_owner_ids: &'a std::collections::HashMap<String, i64>,
